Revert "test"
This reverts commit b4849f6548e555d290eba65941d71cfeb371837d.

+++ cmd/exec.go
@@ -121,7 +121,7 @@ func doExec(ctx *cli.Context) error {
 			log.Warn().Msg("detaching process but pid-file value is empty")
 			return nil
 		}
-		return createPidFile(pidFile, int64(pid))
+		return createPidFile(pidFile, pid)
 	}
 
 	exitStatus, err := c.RunCommandStatus(procArgs, attachOpts)
